refactor(gotrue): move custom provider types to their own file and use enum name
Moves the custom OAuth provider admin types out of types.dart into
custom_oauth_provider.dart, and drops the redundant CustomProviderType.value
in favour of the enum's built-in name.
